backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.135.2",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]
